Created upgrades script for list. Updated shop script. Can purchase upgrades with 1, 2, 3, 4 keys.

+++ Assets/Scripts/Upgrades.cs
@@ -12,9 +12,9 @@ public class Upgrades : IComparable<Upgrades>
     public string name;
     public int level;
     public int cost;
-    public int upgradeAmount;
+    public float upgradeAmount;
 
-    public Upgrades(int newType, string newName, int newLevel, int newCost, int newUpgradeAmount)
+    public Upgrades(int newType, string newName, int newLevel, int newCost, float newUpgradeAmount)
     {
         // button type, upgrade name, weapon level--still upgradable?, cost, upgrade amount
         type = newType;
